pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "molquiz"
version = "0.1.0"
description = "Telegram bot for training organic nomenclature."
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
  "aiogram>=3.13,<4",
  "alembic>=1.13,<2",
  "asyncpg>=0.29,<0.30",
  "fastapi>=0.115,<1",
  "greenlet>=3.1,<4",
  "httpx>=0.27,<1",
  "orjson>=3.10,<4",
  "pillow>=11,<12",
  "prometheus-client>=0.20,<1",
  "pydantic>=2.8,<3",
  "pydantic-settings>=2.4,<3",
  "python-dotenv>=1.0,<2",
  "pyyaml>=6.0,<7",
  "rdkit>=2024.9.1",
  "redis>=5.0,<6",
  "sqlalchemy>=2.0,<3",
  "structlog>=24.4,<25",
  "tenacity>=9.0,<10",
  "typer>=0.12,<1",
  "uvicorn[standard]>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20,<1",
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.24,<1",
  "pytest-cov>=5,<6",
  "respx>=0.22,<1",
  "ruff>=0.6,<1",
]

[project.scripts]
molquiz-web = "molquiz.main:run"
molquiz-dev = "molquiz.dev_polling:run"
molquiz-worker = "molquiz.worker:run"
molquiz-import-pubchem = "molquiz.cli.import_pubchem:main"
molquiz-seed-demo = "molquiz.cli.seed_demo:main"
molquiz-seed-rational = "molquiz.cli.seed_rational:main"
molquiz-generate-depictions = "molquiz.cli.generate_depictions:main"
molquiz-review-export = "molquiz.cli.review_export:main"
molquiz-review-apply = "molquiz.cli.review_apply:main"
molquiz-publish-ready = "molquiz.cli.publish_ready:main"

[tool.hatch.build.targets.wheel]
packages = ["src/molquiz"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
